fix: move tests to standalone minimize_switch_statements and enable optimization in tdz
